models/bge_embeddings/requirements.txt
onnxruntime-gpu==1.16.3
transformers>=4.35.0
